# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Cash Flow Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses/{courseId}/modules:
    get:
      operationId: listModules
      tags: [modules]
      summary: List modules for a course
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Module"
    post:
      operationId: createModule
      tags: [modules]
      summary: Create a module (admin only)
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "201":
          description: Module created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
  /modules/{moduleId}:
    put:
      operationId: updateModule
      tags: [modules]
      summary: Update a module (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModuleBody"
      responses:
        "200":
          description: Module updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
    delete:
      operationId: deleteModule
      tags: [modules]
      summary: Delete a module (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
  /modules/{moduleId}/lessons:
    get:
      operationId: listLessons
      tags: [lessons]
      summary: List lessons for a module
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of lessons
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lesson"
    post:
      operationId: createLesson
      tags: [lessons]
      summary: Create a lesson (admin only)
      parameters:
        - name: moduleId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "201":
          description: Lesson created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
  /lessons/{lessonId}:
    get:
      operationId: getLesson
      tags: [lessons]
      summary: Get a lesson
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lesson detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    put:
      operationId: updateLesson
      tags: [lessons]
      summary: Update a lesson (admin only)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLessonBody"
      responses:
        "200":
          description: Lesson updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lesson"
    delete:
      operationId: deleteLesson
      tags: [lessons]
      summary: Delete a lesson (admin only)
      parameters:
        - name: lessonId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
  /users/{userId}:
    put:
      operationId: updateUser
      tags: [users]
      summary: Update user role (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user (admin only)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [student, admin]
        hasPaid:
          type: boolean
        createdAt:
          type: string
      required: [id, email, name, role, hasPaid, createdAt]
...
    UpdateUserBody:
      type: object
      properties:
        role:
          type: string
          enum: [student, admin]
        hasPaid:
          type: boolean
      required: []
...
    Module:
      type: object
      properties:
        id:
          type: integer
        courseId:
          type: integer
        title:
          type: string
        description:
          type: string
        order:
          type: integer
        createdAt:
          type: string
      required: [id, courseId, title, description, order, createdAt]
...
    CreateModuleBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        order:
          type: integer
      required: [title, description, order]
    Lesson:
      type: object
      properties:
        id:
          type: integer
        moduleId:
          type: integer
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        videoType:
          type: string
          enum: [youtube, vimeo, upload, embed]
        duration:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
        createdAt:
          type: string
      required: [id, moduleId, title, description, videoUrl, videoType, duration, order, isFree, createdAt]
    CreateLessonBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        videoUrl:
          type: string
        videoType:
          type: string
          enum: [youtube, vimeo, upload, embed]
        duration:
          type: integer
        order:
          type: integer
        isFree:
          type: boolean
      required: [title, description, videoUrl, videoType, duration, order, isFree]
...